+++ src/classes/models/game.class.js
@@ -30,8 +30,6 @@ class Game {
       user.init();
       clearUserMonsterData(user.id);
   });
-
-    // 게임 초기화
   }
 
 
@@ -39,10 +37,6 @@ class Game {
   addUser(user) {
     this.users.push(user);
     user.gameId = this.id;
-
-    if (this.users.length >= config.game.maxPlayer) {
-      // throw new Error('게임 인원이 가득 차 참가하실 수 없습니다.');
-    }
 
     if (this.users.length === config.game.maxPlayer) {
       setTimeout(() => {
@@ -129,8 +123,6 @@ class Game {
       user.socket.write(startPacket);
     });
 
-    console.log("GameStart");
-
     return true;
   }
 
@@ -180,9 +172,7 @@ class Game {
         if (newPos.y > 380) {
           newPos.y = 380;
         }
-        // console.log(
-        //   `(${i}, ${j}) => realAngle: ${realAngle}, newPos: (${newPos.x}, ${newPos.y})`,
-        // );
+
         // endPosition에 도달하거나 초과할 때 강제로 마지막 위치를 맞춤
         if (newPos.x >= endPosition.x) {
           path.push({ x: endPosition.x, y: endPosition.y });
