+++ src/constants/game.js
@@ -3,9 +3,6 @@ export const MAX_PLAYER = 2;
 export const GAME_STATE_WAITING = 'waiting';
 export const GAME_STATE_PLAYING = 'playing';
 
-export const CANVAS_WIDTH = 1300;
-export const CANVAS_HEIGH = 360;
-
 export const INIT_BASE_DATA = {
   hp: 50,
   maxHp: 50,
