+++ src/handlers/index.js
@@ -2,14 +2,14 @@ import { HANDLER_IDS } from '../constants/handlerIds.js';
 import { PACKET_TYPE_NAMES } from '../constants/packetTypes.js';
 import { CustomError } from '../utils/error/customError.js';
 import { ErrorCodes } from '../utils/error/errorCodes.js';
-import loginHandler from './login.handler.js';
-import monsterAttackBaseHandler from './monsterAttackBase.handler.js';
-import gameEndHandler from './gameEnd.handler.js';
-import registerHandler from './register.handler.js';
-import { towerPurchaseHandler, towerAttackHandler } from './tower.handler.js';
+import loginHandler from './auth/login.handler.js';
+import monsterAttackBaseHandler from './game/monsterAttackBase.handler.js';
+import gameEndHandler from './game/gameEnd.handler.js';
+import registerHandler from './auth/register.handler.js';
+import { towerPurchaseHandler, towerAttackHandler } from './game/tower.handler.js';
 import matchingGameHandler from './game/matchingGame.handler.js';
-import spawnMonsterHandler from './spawnMonster.handler.js';
-import deathMonsterHandler from './monsterDeath.handler.js';
+import spawnMonsterHandler from './game/spawnMonster.handler.js';
+import deathMonsterHandler from './game/monsterDeath.handler.js';
 
 const handlers = {
   [HANDLER_IDS.REGISTER]: {
